==> notebooks/1_Genius_API_Gathering_Data.ipynb <==
import pandas as pd
import lyricsgenius
import json

# 1. Define Lists of Artists to Use from Billboard Top 100

# using billboard to define top artists of 2020-2021
artists_1 = ['Post Malone','The Weeknd','Roddy Ricch']

artists_2 = ['Rod Wave','Doja Cat','Travis Scott','Morgan Wallen','Lizzo','Future','Ariana Grande','Eminem']

artists_3 = ['Selena Gomez','Camila Cabello','Lady Gaga','Khalid','Lil Mosey','SAINt JHN','Trevor Daniel']

artists_4 = ['Trippie Redd','Lil Nas X','Moneybagg Yo','Mac Miller','XXXTENTACION','DaBaby']

artists_5 = ['NLE Choppa','Lil Wayne','Lee Brice','Demi Lovato','Kodak Black','Surfaces','Lil Tecca','Lil Tjay']

artists_6 = ['Lana Del Rey','Olivia Rodrigo','Ava Max','The Kid LAROI','AJR','Madison Beer','Florida Georgia Line']

artists_7 = ['Conan Gray','girl in red','Wallows','beabadoobee','FINNEAS','Hillsong Worship']

artists_8 = ['Luke Combs','Lil Uzi Vert','Justin Bieber','Lewis Capaldi','YoungBoy Never Broke Again','Bad Bunny']

artists_9 = ['Old Dominion','Luke Bryan','Lil Durk','Fleetwood Mac','Thomas Rhett','Tory Lanez','NF','Jason Derulo']

artists_10 = ['Noah Cyrus','Frank Ocean','Wiz Khalifa','Katy Perry','twenty one pilots','DJ Khaled','Brent Faiyaz']

artists_11 = ['SZA','Cardi B','A Boogie Wit da Hoodie','blackbear','Sam Hunt','Elevation Worship','Upchurch']

artists_12 = ['Drake','Juice WRLD','Lil Baby','Harry Styles','Taylor Swift','Pop Smoke','Billie Eilish','Dua Lipa']

artists_13 = ['Lauren Daigle','Gucci Mane','Glass Animals','Russ','Carrie Underwood','Childish Gambino']

artists_14 = ['Maroon 5','Young Thug','Kane Brown','Chris Brown','Jonas Brothers','Maren Morris','Logic']

artists_15 = ['Jack Harlow','Jason Aldean','Summer Walker','Blake Shelton','Arizona Zervas']

artists_16 = ['Shawn Mendes','James Arthur','Machine Gun Kelly','Don Toliver','YNW Melly','Megan Thee Stallion']

artists_17 = ['Jon Pardi','Miranda Lambert','Chris Stapleton','Sam Smith','Kenny Chesney']

artists_18 = ['Omar Apollo','MorMor','Caamp','Nicki Minaj','Miley Cyrus','Imagine Dragons','24kGoldn','Tate McRae']

artists_19 = ['Playboi Carti','Ashe','Internet Money','Halsey','Queen','Bruno Mars']

artists_20 = ['Saweetie','Nick Jonas','Kendrick Lamar','BTS']

artists_21 = ['Gunna','21 Savage','Tones and I','Polo G','Ed Sheeran']

#setting up token
genius = lyricsgenius.Genius("MT1p0vkqoVRXzFH4kdK3LQPOgGe12XN-0-f0hNz_1H4wtRcn01kMYou9mW-03VZC")

# 2. Set Up Empty DF's to Append To

# appending features to artist df

artist_info = {
    'artist' : [''],
    'description' : [''],
    'images_url' : [''],
    'followers_count' : ['']
    }

artist_df = pd.DataFrame(artist_info)

song_info = {
    'song_name' : [''],
    'artist' : [''],
    'lyrics' : [''],
    'description' : [''],
    'accepted_annotations' : [''],
    'contributors' : [''],
    'pageviews' : ['']
        }

songs_df = pd.DataFrame(song_info)

# 3. Define a Function That Will Add Top 5 Songs For a List of Artists

# This function takes a list of artists and adds them to my df's
def add_artists(list_artists):
    
    # global so it can use defined variables above
    global artist_df, songs_df
    
    # looping through all artists
    for art in list_artists:
        # read in top 5 songs for the artist
        artist = genius.search_artist(art, max_songs = 7)
        # save lyrics and song info to json
        artist.save_lyrics()
        # making json file path as string
        temp_art = art.replace(" ", "")
        temp_art = "Lyrics_" + temp_art + ".json"
        # reading in json
        with open(temp_art) as json_data:
            data = json.load(json_data)

        # appending features to artist df
        artist_info_temp = {
        'artist' : [data['name']],
        'description' : [data['description']['plain'].replace('\n',' ').replace("\'","'")],
        'images_url' : [data['image_url']],
        'followers_count' : [data['followers_count']]
        }
        artist_df_temp = pd.DataFrame(artist_info_temp)
        artist_df = artist_df.append(artist_df_temp, ignore_index=True)

        # appending features to song df
        for i in range(7):
            song_info_temp = {
            'song_name' : [data['songs'][i]['title']],
            'artist' : [data['songs'][i]['artist']],
            'lyrics' : [data['songs'][i]['lyrics'].replace('\n',' ').replace("\'","'")],
            'description' : [data['songs'][i]['description']['plain'].replace('\n',' ').replace("\'","'")],
            'accepted_annotations' : [data['songs'][i]['stats']['accepted_annotations']],
            'contributors' : [data['songs'][i]['stats']['contributors']],
            'pageviews' : [data['songs'][i]['stats']['pageviews']]
            }
            songs_df_temp = pd.DataFrame(song_info_temp)
            songs_df = songs_df.append(songs_df_temp, ignore_index=True)

# 4. Add Artists From Billboard Lists

add_artists(artists_1)

add_artists(artists_2)

add_artists(artists_3)

add_artists(artists_4)

add_artists(artists_5)

add_artists(artists_6)

add_artists(artists_7)

add_artists(artists_8)

add_artists(artists_9)

add_artists(artists_10)

add_artists(artists_11)

add_artists(artists_12)

add_artists(artists_13)

add_artists(artists_14)

add_artists(artists_15)

add_artists(artists_16)

add_artists(artists_17)

add_artists(artists_18)

add_artists(artists_19)

add_artists(artists_20)

add_artists(artists_21)

# 5. Export Both DF's as CSV's

# dropping duplicate rows
artist_df1 = artist_df.drop_duplicates()
songs_df1 = songs_df.drop_duplicates()

#saving to csv
artist_df1.to_csv('../data/artist.csv', index=False)
songs_df1.to_csv('../data/songs.csv', index=False)
