==> notebooks/2_Data_Cleaning_and_Eda.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
import re
import warnings
warnings.filterwarnings('ignore')
pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)

artists = pd.read_csv('../data/artist.csv')
songs = pd.read_csv('../data/songs.csv')

artists = artists.iloc[1: , :]
songs = songs.iloc[1: , :]

# Artists

artists.head()

# Songs

songs.head()

# Data Cleaning: Artist Gender

# Function to Count How Many Gendered Words Appear in Description 
def count_male(row):
    num_male = 0
    male_words = [' his ',' he ',' him ',' himself ',' man ',' actor ']
    for word in male_words:
        num_male += row['description'].lower().count(word)
    return num_male

def count_female(row):
    num_female = 0
    female_words = [' her ',' hers ',' she ',' herself ',' woman ',' actress ']
    for word in female_words:
        num_female += row['description'].lower().count(word)
    return num_female

artists['male_count'] = artists.apply(count_male, axis = 1)
artists['female_count'] = artists.apply(count_female, axis = 1)

# Function to Get Artist Gender from Gender Word Counts
def get_gender(row):
    if row['male_count'] > row['female_count']:
        return "Male"
    elif row['female_count'] > row['male_count']:
        return "Female"
    else:
        return None
    
artists['gender'] = artists.apply(get_gender, axis = 1)

# Catching missed genders
missed = {'Gabby Barrett':'Female', 'Sam Smith':'Non Binary', 'Demi Lovato':'Non Binary'}
art_miss = ['Gabby Barrett','Sam Smith','Demi Lovato']
artists["missed"] = artists["artist"].map(missed)

#Correcting mis-gendered artists
def correct_gender(row):
    if row['artist'] in art_miss:
        return row['missed']
    else:
        return row['gender']
    
artists['gender'] = artists.apply(correct_gender, axis = 1)
artists = artists.drop(columns=['male_count', 'female_count','missed'])

import plotly.express as px
gen = pd.DataFrame(artists['gender'].value_counts()).reset_index()
gen.head()
fig = px.bar(gen, x='index', y='gender',color = 'index',color_discrete_sequence=['#FFD700','black','#FF6347'])
fig.update_layout(
    autosize=False,
    width=500,
    height=400)

fig.show()

# Data Cleaning: Artist Age

Most birth years are recorded like this: '(b. December 6, 1995)' towards the begining of the description

I am looking for the first occurance of the pattern ' ####)' to get artist birth year

# getting first year that occirs in text and checking that its not too recent
def birth_year(row):
    years_list = re.findall(r"\d\d\d\d", row['description'])
    if not years_list:
        return None
    else:
        if int(years_list[0]) < 2004:
            return int(years_list[0])
        else:
            return None
    
artists['birth_year'] = artists.apply(birth_year, axis = 1)

# Catching missed birth years
missed_year = {'Taylor Swift': 1989,'Luke Combs': 1990,
               'Chris Brown': 1989, 'Selena Gomez': 1992,
               'Maren Morris': 1990,'Halsey': 1994,
               'Tones and I': 2000,'Ed Sheeran': 1991,
               'Gabby Barrett': 2000,'Lady Gaga': 1986,
               'SAINt JHN': 1986,'Trevor Daniel': 1994,
               'Jason Aldean': 1977,'Arizona Zervas': 1995,
               '\u200bblackbear': 1990,'Moneybagg Yo': 1999,
               'Machine Gun Kelly': 1990,'NF': 1991,'Kenny Chesney': 1968,
               'Lana Del Rey': 1985,'Jason Derulo': 1989,'Ashe': 1993}

year_miss = ['Taylor Swift','Luke Combs','Chris Brown','Selena Gomez',
             'Maren Morris','Halsey','Tones and I','Ed Sheeran',
             'Gabby Barrett','Lady Gaga','SAINt JHN','Trevor Daniel',
             'Jason Aldean','Arizona Zervas','\u200bblackbear','Moneybagg Yo',
             'Machine Gun Kelly','NF','Kenny Chesney',
             'Lana Del Rey','Jason Derulo','Ashe']

artists["missed_year"] = artists["artist"].map(missed_year)

#Correcting mis-gendered artists
def correct_year(row):
    if row['artist'] in year_miss:
        return row['missed_year']
    else:
        return row['birth_year']
    
artists['birth_year'] = artists.apply(correct_year, axis = 1)
artists = artists.drop(columns=['missed_year'])

#Getting Decade Born
def decade(row):
    if row['birth_year'] in [1960,1961,1962,1963,1964,1965,1966,1967,1968,1969]:
        return "1960s"
    elif row['birth_year'] in [1970,1971,1972,1973,1974,1975,1976,1977,1978,1979]:
        return "1970s"
    elif row['birth_year'] in [1980,1981,1982,1983,1984,1985,1986,1987,1988,1989]:
        return "1980s"
    elif row['birth_year'] in [1990,1991,1992,1993,1994,1995,1996,1997,1998,1999]:
        return "1990s"
    elif row['birth_year'] in [2000,2001,2002,2003,2004]:
        return "2000s"
    else:
        return None
    
artists['birth_decade'] = artists.apply(decade, axis = 1)

# making buckets by 5 years
def five_years(row):
    if row['birth_year'] in [1960,1961,1962,1963,1964]:
        return "1960-1964"
    elif row['birth_year'] in [1965,1966,1967,1968,1969]:
        return "1965-1969"
    elif row['birth_year'] in [1970,1971,1972,1973,1974]:
        return "1970-1974"
    elif row['birth_year'] in [1975,1976,1977,1978,1979]:
        return "1975-1979"
    elif row['birth_year'] in [1980,1981,1982,1983,1984]:
        return "1980-1984"
    elif row['birth_year'] in [1985,1986,1987,1988,1989]:
        return "1985-1989"
    elif row['birth_year'] in [1990,1991,1992,1993,1994]:
        return "1990-1994"
    elif row['birth_year'] in [1995,1996,1997,1998,1999]:
        return "1995-1999"
    elif row['birth_year'] in [2000,2001,2002,2003,2004]:
        return "2000-2004"
    else:
        return None
    
artists['birth_5_yr'] = artists.apply(five_years, axis = 1)

import plotly.express as px

fig = px.histogram(artists[artists['birth_year']>1955], x="birth_year", nbins = 21,
                   labels={'birth_year':'Artist Birth Year', 'Count':'Count'}, 
                   color_discrete_sequence=['#FFD700'],
                   opacity=0.8)
fig.update_layout(
    autosize=False,
    width=500,
    height=400)
fig.show()

# Exporting Cleaned Data

artists.to_csv('../data/artist_cleaned.csv', index=False)
songs.to_csv('../data/songs_cleaned.csv', index=False)
