==> notebooks/6_Generate_Pop_Song.ipynb <==
import warnings
warnings.filterwarnings('ignore')
import pandas as pd
pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)
import matplotlib.pyplot as plt
%matplotlib inline
import matplotlib
from collections import Counter
from random import randrange
from profanity_check import predict, predict_prob

songs = pd.read_csv('../data/songs_after_round_3_cleaning.csv')

# Build Lyrics Corpus

lyrics_corpus = ""

for l in list(songs['cleaned_lyrics']):
    lyrics_corpus += l

lyrics_list = lyrics_corpus.split()

lyrics_list = [s.replace("[", "") for s in lyrics_list]
lyrics_list = [s.replace("]", "") for s in lyrics_list]

# Get Probabilty Word Follows Another Word to Generate Lyrics

def most_frequent(List):
    occurence_count = Counter(List)
    return occurence_count.most_common(10)

dirty = ['fuck','ass','shit','bitch','cunt','nigga','nigger','niggas','niggers','asshole','penis','bitches']

def create_lyrics(start_word, length):
    
    new_word = start_word
    lyrics = start_word + ' '
    
    for i in range(length):
        
        following_words = []
        
        for i in range(len(lyrics_list)):
            if (lyrics_list[i] == new_word) and (i != len(lyrics_list) -1):
                following_words.append(lyrics_list[i+1])
    
        if len(set(following_words)) > 40:
            fol = most_frequent(following_words)[randrange(8)][0]
            if fol in dirty:
                fol = most_frequent(following_words)[randrange(7)][0]
            if fol in dirty:
                fol = most_frequent(following_words)[randrange(7)][0]
            lyrics += fol + " "
            new_word = fol
            
        else:
                lyrics += str(most_frequent(following_words)[0][0]) + " "
                new_word = most_frequent(following_words)[0][0]
                
       
            
        
    return(lyrics)

create_lyrics('friends', 12)

create_lyrics('a', 12)

create_lyrics('and', 12)

create_lyrics('he', 12)

# Write Song Function

# list of first words in songs
starter = ['i','im','you','dont','every',
           'and','oh','yeah','youre','baby','so','she',
           'he','they','this','well','we','hey','if','how','hold','wanting','just','medicine','yesterday',
            'night','morning','sun','friend',
            'first','time','anxiety','loving']

def write_song(row_length):
    l = []
    for j in range(row_length):
        l.append(create_lyrics(starter[randrange(len(starter))], randrange(5,15)))
    return l

for thing in write_song(15):
    print(thing)

for thing in write_song(7):
    print(thing)

# Adding Pegasus Text Summarization

from transformers import PegasusForConditionalGeneration, PegasusTokenizer
import torch
import io
import seaborn as sb
from transformers import pipeline
from transformers import AutoModelWithLMHead, AutoTokenizer

len(lyrics_corpus)

lyrics_corpus[0:500]

model_name = 'google/pegasus-xsum'
torch_device = 'cuda' if torch.cuda.is_available() else 'cpu'
tokenizer = PegasusTokenizer.from_pretrained(model_name)
model = PegasusForConditionalGeneration.from_pretrained(model_name).to(torch_device)

batch = tokenizer.prepare_seq2seq_batch(lyrics_corpus[0:500], truncation=True, padding='longest', return_tensors='pt').to(torch_device)
translated = model.generate(**batch)
tgt_text = tokenizer.batch_decode(translated, skip_special_tokens=True)
print(tgt_text[0])

def write_song(row):
    batch = tokenizer.prepare_seq2seq_batch(row['cleaned_lyrics'], truncation=True, padding='longest', return_tensors='pt').to(torch_device)
    translated = model.generate(**batch)
    tgt_text = tokenizer.batch_decode(translated, skip_special_tokens=True)
    return str(tgt_text[0])

list(songs['cleaned_lyrics'])[0]

songs['pegasus'] = songs.apply(write_song, axis = 1)

#Making pegasus summary and appending into a list
short_summary = []
batch = tokenizer.prepare_seq2seq_batch(lyrics_corpus[0:30], truncation=True, padding='longest', return_tensors='pt').to(torch_device)
translated = model.generate(**batch)
tgt_text = tokenizer.batch_decode(translated, skip_special_tokens=True)
short_summary.append(str(tgt_text[0]))
